Upload interface.py
Uploading this here after a corruption incident. It's an older version of the script which I now need to (almost entirely) rewrite after many, many hours of work.
For context I've rewritten the old script almost entirely, commented it and added logging. The only thing that's left now is one single function and the logging configuration.
Backup your code fellas.